Fix CollectionItem type mismatch: i64 → i32 to match PG INT4
CollectionItem.imageId and collectionId are INT4 (i32) in Postgres,
not BIGINT (i64). The structs incorrectly used i64, causing sqlx type
mismatch errors at runtime. Also removed unnecessary ::int8 casts
from the enrichment query.

- CollectionItemRow: image_id/collection_id → i32
- StreamCollectionRow: collection_id/image_id → i32
- get_max_collection_id: returns i32
- fetch_collections_by_range: takes i32 start/end
- Outbox poller: explicit i32→i64 widening at API boundary

+++ src/pg_sync/outbox_poller.rs
@@ -216,7 +216,7 @@ async fn fetch_and_push_upserts(
             coll_map.entry(id).or_default();
         }
         for c in &collections {
-            coll_map.entry(c.image_id).or_default().push(c.collection_id);
+            coll_map.entry(c.image_id as i64).or_default().push(c.collection_id as i64);
         }
         let mut entries: Vec<(i64, Vec<i64>)> = coll_map.into_iter().collect();
         entries.sort_by_key(|(id, _)| *id); // Deterministic ordering for debugging
